Safari-609.1.11.1/LayoutTests/js/exception-instanceof-expected.txt

Test for error messages for instanceof

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'instanceof' instanceof    'instanceof' threw exception TypeError: Right hand side of instanceof is not an object.
PASS 20 instanceof     'hello'   threw exception TypeError: Right hand side of instanceof is not an object.
PASS 20 instanceof     {}   threw exception TypeError: {} is not a function. (evaluating '20 instanceof     {}').
PASS 20 instanceof     {}.foo  threw exception TypeError: Right hand side of instanceof is not an object.
PASS 20 instanceof     true       threw exception TypeError: Right hand side of instanceof is not an object.
PASS successfullyParsed is true

TEST COMPLETE
